ci: build: Build wheel and upload slsa.dev evidence
## Description

We're working on ensuring all Intel repos are generating https://slsa.dev style evidence. See goto/securereuse for more details. Still in early stages of communicating what we're looking for. However, the gist of it is that we want to ensure all released artifacts have SLSA info in a transparency service where customers can verify it.

Related: https://github.blog/2024-05-02-introducing-artifact-attestations-now-in-public-beta/